tidy up structure
